+++ src/common/hooks/index.js
@@ -6,6 +6,8 @@ export * from './useInvoiceMutation';
 export * from './useInvoiceLine';
 export * from './useInvoiceLineMutation';
 export * from './useOrderLine';
+export * from './useOrderLines';
+export * from './useOrders';
 export * from './useVoucherById';
 export * from './useVoucherByInvoiceId';
 export * from './useVoucherLines';