+++ src/invoices/InvoiceForm/InvoiceFormContainer.js
@@ -16,6 +16,7 @@ import {
   LIMIT_MAX,
   sourceValues,
   useModalToggle,
+  useOrganization,
   useShowCallout,
   VENDORS_API,
 } from '@folio/stripes-acq-components';
@@ -31,11 +32,17 @@ import {
   invoiceDocumentsResource,
   invoiceResource,
   invoicesResource,
-  VENDOR,
 } from '../../common/resources';
+import {
+  useInvoice,
+  useInvoiceLineMutation,
+  useOrderLines,
+  useOrders,
+} from '../../common/hooks';
 import {
   getSettingsAdjustmentsList,
 } from '../../settings/adjustments/util';
+import { createInvoiceLineFromPOL } from '../InvoiceDetails/utils';
 import InvoiceForm from './InvoiceForm';
 import {
   saveInvoice,
@@ -44,6 +51,8 @@ import {
 import DuplicateInvoiceModal from './DuplicateInvoiceModal/DuplicateInvoiceModal';
 
 export function InvoiceFormContainerComponent({
+  location,
+  history,
   match,
   mutator,
   okapi,
@@ -51,8 +60,25 @@ export function InvoiceFormContainerComponent({
   resources,
   stripes,
 }) {
+  const { params: { id } } = match;
+  const isCreate = !id;
+  const showToast = useShowCallout();
   const [batchGroups, setBatchGroups] = useState();
   const [duplicateInvoices, setDuplicateInvoices] = useState();
+  const orderIds = location?.state?.orderIds;
+  const { invoice, isInvoiceLoading } = useInvoice(id);
+  const { orders, isLoading: isOrdersLoading } = useOrders(orderIds ? [orderIds[0]] : undefined);
+  const { orderLines, isLoading: isOrderLinesLoading } = useOrderLines(orderIds);
+  const invoiceVendorId = isCreate ? orders?.[0]?.vendor : invoice.vendorId;
+  const { organization: invoiceVendor, isLoading: isVendorLoading } = useOrganization(invoiceVendorId);
+  const { mutateInvoiceLine } = useInvoiceLineMutation({
+    onSuccess: () => {
+      return showToast({ messageId: 'ui-invoice.invoiceLine.hasBeenSaved' });
+    },
+    onError: () => {
+      return showToast({ messageId: 'ui-invoice.errors.invoiceLineHasNotBeenSaved', type: 'error' });
+    },
+  });
 
   useEffect(() => {
     setBatchGroups();
@@ -68,9 +94,6 @@ export function InvoiceFormContainerComponent({
 
   const [isNotUniqueOpen, toggleNotUnique] = useModalToggle();
   const [forceSaveValues, setForceSaveValues] = useState();
-  const showToast = useShowCallout();
-  const { params: { id } } = match;
-  const isCreate = !id;
   const invoiceDocuments = get(resources, 'invoiceFormDocuments.records', []);
 
   const saveInvoiceHandler = useCallback(formValues => {
@@ -97,32 +120,58 @@ export function InvoiceFormContainerComponent({
             setDuplicateInvoices(duplicates);
 
             // eslint-disable-next-line prefer-promise-reject-errors
-            return Promise.reject({ validationError: VALIDATION_ERRORS.dublicateInvoice });
+            return Promise.reject({ validationError: VALIDATION_ERRORS.duplicateInvoice });
           }
         });
     }
 
     return validationRequest
       .then(() => saveInvoice(formValues, invoiceDocuments, mutator.invoiceFormInvoices, okapi))
-      .then(savedRecord => {
+      .then(async (savedRecord) => {
         showToast({ messageId: 'ui-invoice.invoice.invoiceHasBeenSaved' });
 
-        setTimeout(() => onCancel(savedRecord.id));
+        if (orderLines?.length) {
+          if (orderLines.length > 1) return history.push(`/invoice/view/${savedRecord.id}/lines-sequence`, { orderIds });
+
+          const data = createInvoiceLineFromPOL(orderLines[0], savedRecord.id, invoiceVendor);
+
+          try {
+            await mutateInvoiceLine({ data });
+          } catch {
+            return onCancel(savedRecord.id);
+          }
+        }
+
+        return onCancel(savedRecord.id);
       })
       .catch(({ validationError }) => {
-        if (validationError === VALIDATION_ERRORS.dublicateInvoice) return toggleNotUnique();
+        if (validationError === VALIDATION_ERRORS.duplicateInvoice) return toggleNotUnique();
 
         return showToast({ messageId: 'ui-invoice.errors.invoiceHasNotBeenSaved', type: 'error' });
       });
+  },
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [forceSaveValues, id, toggleNotUnique, invoiceDocuments, okapi, showToast, onCancel]);
+  [
+    forceSaveValues,
+    id,
+    toggleNotUnique,
+    invoiceDocuments,
+    okapi,
+    showToast,
+    onCancel,
+    orderIds,
+    orderLines,
+    history,
+    mutateInvoiceLine,
+    invoiceVendor,
+  ]);
 
   const allAdjustments = getSettingsAdjustmentsList(get(resources, ['configAdjustments', 'records'], []));
   const alwaysShowAdjustments = getAlwaysShownAdjustmentsList(allAdjustments);
   const batchGroupId = isCreate && (batchGroups?.length === 1) ? batchGroups[0]?.id : undefined;
   const exportToAccounting = alwaysShowAdjustments.some(adj => adj.exportToAccounting);
-  const invoice = !isCreate
-    ? get(resources, ['invoice', 'records', 0], {})
+  const initialInvoice = !isCreate
+    ? invoice
     : {
       chkSubscriptionOverlap: true,
       currency: stripes.currency,
@@ -131,18 +180,16 @@ export function InvoiceFormContainerComponent({
       batchGroupId,
       status: INVOICE_STATUS.open,
       exportToAccounting,
+      vendorId: orders?.[0]?.vendor,
     };
 
-  const invoiceVendor = isCreate ? undefined : get(resources, ['invoiceFormVendor', 'records', 0]);
   const initialValues = {
-    ...invoice,
+    ...initialInvoice,
     documents: invoiceDocuments.filter(invoiceDocument => !invoiceDocument.url),
     links: invoiceDocuments.filter(invoiceDocument => invoiceDocument.url),
   };
 
-  const hasLoaded = (!id && batchGroups) || (
-    get(resources, 'invoice.hasLoaded') && get(resources, 'invoiceFormVendor.hasLoaded') && batchGroups
-  );
+  const hasLoaded = batchGroups && !(isInvoiceLoading || isOrdersLoading || isOrderLinesLoading || isVendorLoading);
 
   if (!hasLoaded) {
     return (
@@ -161,6 +208,7 @@ export function InvoiceFormContainerComponent({
         onSubmit={saveInvoiceHandler}
         onCancel={onCancel}
         batchGroups={batchGroups}
+        isCreateFromOrder={Boolean(orderIds?.length)}
       />
       {
         isNotUniqueOpen && (
@@ -179,7 +227,10 @@ export function InvoiceFormContainerComponent({
 }
 
 InvoiceFormContainerComponent.manifest = Object.freeze({
-  invoice: invoiceResource,
+  invoice: {
+    ...invoiceResource,
+    fetch: false,
+  },
   invoiceFormDocuments: {
     ...invoiceDocumentsResource,
     accumulate: true,
@@ -196,12 +247,13 @@ InvoiceFormContainerComponent.manifest = Object.freeze({
   },
   configAdjustments: CONFIG_ADJUSTMENTS,
   configAddress,
-  invoiceFormVendor: VENDOR,
   batchGroups: batchGroupsResource,
 });
 
 InvoiceFormContainerComponent.propTypes = {
   match: ReactRouterPropTypes.match.isRequired,
+  history: ReactRouterPropTypes.history.isRequired,
+  location: ReactRouterPropTypes.location.isRequired,
   mutator: PropTypes.object.isRequired,
   okapi: PropTypes.object.isRequired,
   onCancel: PropTypes.func,