+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * Error message does not contain the reason of failing invoice line adding. Refs UINV-413.
 * Grant budget permissions. Refs UINV-415.
 * Error handling for approving invoice. Refs UINV-418.
+* Create new invoice from order. Refs UINV-420.
 
 ## [3.1.1](https://github.com/folio-org/ui-invoice/tree/v3.1.1) (2022-03-22)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.1.0...v3.1.1)